frontend/requirements.txt
Flask==1.1.2
python-dotenv==0.15.0
gunicorn==20.0.4
pandas==1.1.4
numpy==1.19.4
matplotlib==3.3.3
